--- a/lib/parse/src/combinator.rs
+++ b/lib/parse/src/combinator.rs
@@ -5,48 +5,36 @@ use nom::{
     IResult, InputLength, Parser,
 };
 
-pub fn separated_or_terminated_list0<I, O, O2, E, F, G>(
-    sep: G,
-    f: F,
-) -> impl FnMut(I) -> IResult<I, Vec<O>, E>
-where
-    I: Clone + InputLength,
-    F: Parser<I, O, E>,
-    G: Parser<I, O2, E> + Clone,
-    E: ParseError<I>,
-{
-    let mut p = separated_list0(sep.clone(), f);
-    let mut term = opt(sep);
+pub fn separated_or_terminated_list0<I: Clone + InputLength, O, S, E: ParseError<I>>(
+    separator: impl Parser<I, S, E> + Clone,
+    element: impl Parser<I, O, E>,
+) -> impl FnMut(I) -> IResult<I, Vec<O>, E> {
+    let mut list = separated_list0(separator.clone(), element);
+    let mut end = opt(separator);
 
     move |input: I| {
-        let (input, xs) = p(input)?;
+        let (input, xs) = list(input)?;
 
         Ok(if xs.is_empty() {
             (input, xs)
         } else {
-            let (input, _) = term(input)?;
+            let (input, _) = end(input)?;
 
             (input, xs)
         })
     }
 }
 
-pub fn separated_or_terminated_list1<I, O, O2, E, F, G>(
-    sep: G,
-    f: F,
-) -> impl FnMut(I) -> IResult<I, Vec<O>, E>
-where
-    I: Clone + InputLength,
-    F: Parser<I, O, E>,
-    G: Parser<I, O2, E> + Clone,
-    E: ParseError<I>,
-{
-    let mut p = separated_list1(sep.clone(), f);
-    let mut term = opt(sep);
+pub fn separated_or_terminated_list1<I: Clone + InputLength, O, S, E: ParseError<I>>(
+    separator: impl Parser<I, S, E> + Clone,
+    element: impl Parser<I, O, E>,
+) -> impl FnMut(I) -> IResult<I, Vec<O>, E> {
+    let mut list = separated_list1(separator.clone(), element);
+    let mut end = opt(separator);
 
     move |input: I| {
-        let (input, xs) = p(input)?;
-        let (input, _) = term(input)?;
+        let (input, xs) = list(input)?;
+        let (input, _) = end(input)?;
 
         Ok((input, xs))
     }