--- a/lib/parse/src/combinator.rs
+++ b/lib/parse/src/combinator.rs
@@ -39,3 +39,88 @@ pub fn separated_or_terminated_list1<I: Clone + InputLength, O, S, E: ParseError
         Ok((input, xs))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::error::NomError;
+    use crate::input::input;
+    use nom::bytes::complete::tag;
+    use nom::combinator::all_consuming;
+    use std::str;
+
+    mod separated_or_terminated_list0 {
+        use super::*;
+
+        #[test]
+        fn parse() {
+            for (source, expected) in [
+                ("", &[] as &[&str]),
+                ("a", &["a"]),
+                ("a,a", &["a", "a"]),
+                ("a,a,", &["a", "a"]),
+            ] {
+                assert_eq!(
+                    separated_or_terminated_list0(
+                        |input| tag::<_, _, NomError>(",")(input),
+                        tag("a")
+                    )(input(source, ""))
+                    .unwrap()
+                    .1
+                    .iter()
+                    .map(|span| str::from_utf8(span.as_bytes()).unwrap())
+                    .collect::<Vec<_>>(),
+                    expected
+                );
+            }
+        }
+
+        #[test]
+        fn fail_to_parse() {
+            for source in [",", "a,,"] {
+                assert!(all_consuming(separated_or_terminated_list0(
+                    |input| tag::<_, _, NomError>(",")(input),
+                    tag("a")
+                ))(input(source, ""))
+                .is_err());
+            }
+        }
+    }
+
+    mod separated_or_terminated_list1 {
+        use super::*;
+
+        #[test]
+        fn parse() {
+            for (source, expected) in [
+                ("a", &["a"] as &[&str]),
+                ("a,a", &["a", "a"]),
+                ("a,a,", &["a", "a"]),
+            ] {
+                assert_eq!(
+                    separated_or_terminated_list1(
+                        |input| tag::<_, _, NomError>(",")(input),
+                        tag("a")
+                    )(input(source, ""))
+                    .unwrap()
+                    .1
+                    .iter()
+                    .map(|span| str::from_utf8(span.as_bytes()).unwrap())
+                    .collect::<Vec<_>>(),
+                    expected
+                );
+            }
+        }
+
+        #[test]
+        fn fail_to_parse() {
+            for source in ["", ",", "a,,"] {
+                assert!(all_consuming(separated_or_terminated_list1(
+                    |input| tag::<_, _, NomError>(",")(input),
+                    tag("a")
+                ))(input(source, ""))
+                .is_err());
+            }
+        }
+    }
+}