--- a/lib/parse/src/error.rs
+++ b/lib/parse/src/error.rs
@@ -1,6 +1,6 @@
-use crate::input::Input;
-use nom::error::VerboseError;
-use position::{test::PositionFake, Position};
+use crate::input::{position, Input};
+use nom::error::{VerboseError, VerboseErrorKind};
+use position::Position;
 use std::{error::Error, fmt, fmt::Display};
 
 pub type NomError<'a> = VerboseError<Input<'a>>;
@@ -8,45 +8,46 @@ pub type NomError<'a> = VerboseError<Input<'a>>;
 #[derive(Debug, PartialEq, Eq)]
 pub struct ParseError {
     message: String,
-    expected: Vec<String>,
     position: Position,
 }
 
 impl ParseError {
-    pub fn new<'a>(_source: &str, _path: &str, error: nom::Err<NomError<'a>>) -> Self {
+    pub fn new<'a>(source: &str, path: &str, error: nom::Err<NomError<'a>>) -> Self {
+        match error {
+            nom::Err::Incomplete(_) => Self::unexpected_end(source, path),
+            nom::Err::Error(error) | nom::Err::Failure(error) => {
+                if let Some((input, kind)) = error.errors.last() {
+                    Self {
+                        message: match kind {
+                            VerboseErrorKind::Context(context) => {
+                                format!("failed to parse {}", context)
+                            }
+                            VerboseErrorKind::Char(character) => {
+                                format!("letter '{}' expected", character)
+                            }
+                            VerboseErrorKind::Nom(_) => "failed to parse".into(),
+                        },
+                        position: position(*input),
+                    }
+                } else {
+                    Self::unexpected_end(source, path)
+                }
+            }
+        }
+    }
+
+    fn unexpected_end(source: &str, path: &str) -> Self {
+        let lines = source.split('\n').collect::<Vec<_>>();
+        let line = lines
+            .iter()
+            .rev()
+            .find(|string| !string.is_empty())
+            .map(|string| string.to_string())
+            .unwrap_or_default();
+
         Self {
-            message: format!("{}", error),
-            expected: vec![],
-            position: Position::fake(),
-            // TODO
-            // message: errors
-            //     .errors
-            //     .iter()
-            //     .rev()
-            //     .find_map(|error| match error {
-            //         easy::Error::Expected(_) => None,
-            //         easy::Error::Message(info) => Some(info.to_string()),
-            //         easy::Error::Other(error) => Some(error.to_string()),
-            //         easy::Error::Unexpected(info) => Some(format!("unexpected {}", info)),
-            //     })
-            //     .unwrap_or_else(|| "failed to parse module".into()),
-            // expected: errors
-            //     .errors
-            //     .iter()
-            //     .filter_map(|error| match error {
-            //         easy::Error::Expected(info) => Some(info.to_string()),
-            //         _ => None,
-            //     })
-            //     .collect(),
-            // position: Position::new(
-            //     path,
-            //     errors.position.line as usize,
-            //     errors.position.column as usize,
-            //     source
-            //         .split('\n')
-            //         .nth(errors.position.line as usize - 1)
-            //         .unwrap_or_default(),
-            // ),
+            message: "unexpected end of source".into(),
+            position: Position::new(path, lines.len(), line.len(), line),
         }
     }
 }
@@ -58,19 +59,7 @@ impl Display for ParseError {
         write!(
             formatter,
             "{}",
-            [
-                Some(self.message.clone()),
-                if self.expected.is_empty() {
-                    None
-                } else {
-                    Some(format!("expected: {}", self.expected.join(", ")))
-                },
-                Some(self.position.to_string()),
-            ]
-            .into_iter()
-            .flatten()
-            .collect::<Vec<_>>()
-            .join("\n"),
+            [self.message.as_str(), &self.position.to_string()].join("\n"),
         )
     }
 }