--- a/lib/parse/src/lib.rs
+++ b/lib/parse/src/lib.rs
@@ -1,6 +1,7 @@
 mod error;
 mod operations;
 mod parsers;
+mod parsers_v2;
 mod stream;
 
 use ast::Comment;