--- a/lib/parse/src/lib.rs
+++ b/lib/parse/src/lib.rs
@@ -1,3 +1,4 @@
+mod combinator;
 mod error;
 mod operations;
 mod parsers;